added class "DemoTurners".
this class has got a main class to make an object out of each classes of the project and then iterate through each of them by a for loop.
in this class I handled all the objects in an array as so they are stored in the array implement the Turner interface, which means that they can be treated as Turner objects and their "turn()" method can be called in a polymorphic way, regardless of their actual concrete type.